move NativeAppStartEventProcessor to own file
